requirements.txt
absl-py==0.13.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
cycler==0.10.0
decorator==4.4.2
google-auth==1.33.1
google-auth-oauthlib==0.4.4
grpcio==1.39.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.0
Pillow==8.3.1
prettytable==2.1.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyecharts==1.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
simplejson==3.17.3
six==1.16.0
soupsieve==2.2.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpool==1.3.2
threadpoolctl==2.2.0
tifffile==2021.7.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
